Remove `blockOwnerDeletion`
Seems we are not able to set this field, causing a lot of warnings to
show up in the logs. Further, according the Kubernetes documentation[0],
the field is automatically set to true, when we add any
`metadata.ownerReferences` information to the object. As a result, we
don't actually need to set it manually.

[0]: https://kubernetes.io/docs/concepts/overview/working-with-objects/owners-dependents/#owner-references-in-object-specifications

Signed-off-by: Jacob Baungard Hansen <jacobbaungard@redhat.com>